requirements.txt
amqp==2.5.2
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.2.7
beautifulsoup4==4.9.0
billiard==3.6.3.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.5
django-annoying==0.10.6
django-appconf==1.0.4
django-bootstrap3==12.1.0
django-bootstrap4==1.1.1
django-celery==3.3.1
django-crispy-forms==1.9.0
django-custom-500==0.1.1
django-environment==0.3.8
django-filter==2.2.0
django-rest-swagger==2.2.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
drf-yasg==1.17.1
environ==1.0
filelock==3.0.12
foreman==0.9.7
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
kombu==4.6.8
Markdown==3.2.2
MarkupSafe==1.1.1
numpy==1.18.4
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pbr==5.4.5
Pillow==7.1.2
psycopg2-binary==2.8.5
PyJWT==1.7.1
pyparsing==2.4.7
python3-openid==3.1.0
pytz==2020.1
redis==3.5.1
requests==2.23.0
requests-futures==1.0.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
simplejson==3.17.0
six==1.14.0
sorl-thumbnail==12.6.3
soupsieve==2.0
sqlparse==0.3.1
stevedore==1.32.0
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.20
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
whitenoise==5.0.1
